Move data modules to separate folder
